[circuit]: Replace the stored-value in the 'MAX_EMAIL_HEADER_LENGTH' and 'EMAIL_LARGE_MAX_HEADER_LENGTH' with 2048 + Replace the stored-value in the 'MAX_EMAIL_BODY_LENGTH' and 'EMAIL_LARGE_MAX_BODY_LENGTH' with 64000

+++ app/lib/circuits/openbands.ts
@@ -84,12 +84,12 @@ export const OPENBANDS_CIRCUIT_HELPER = {
 
     // @dev - Input data for an Email verification /w ZKEmail.nr
     const headerUint8Array = new Uint8Array(MAX_HEADER_LENGTH);
-    headerUint8Array.set(header.storage as Uint8Array);
+    headerUint8Array.set(header.storage);
     //headerUint8Array.set(Uint8Array.from(new TextEncoder().encode(header)));
     console.log(`headerUint8Array: ${headerUint8Array}`);
 
     const bodyUint8Array = new Uint8Array(MAX_BODY_LENGTH);
-    bodyUint8Array.set(body.storage as Uint8Array);
+    bodyUint8Array.set(body.storage);
     //bodyUint8Array.set(Uint8Array.from(new TextEncoder().encode(body)));
     console.log(`bodyUint8Array: ${bodyUint8Array}`);
